backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
celery[redis]
django-celery-beat
aiohttp
python-dotenv
